chore: FSD 구조에 맞게 UI 컴포넌트 파일 이동
